downloadData function reports if it did update a file (rc=1) or not (rc=0)

+++ bin/fetch_clamav.sh
@@ -74,11 +74,13 @@ fi
 
 . $SRCDIR/lib/updates/download_files.sh
 
-downloadDatas "$VARDIR/spool/clamav/" "clamav3" $randomize "clamav" "\|main.cvd\|bytecode.cvd\|daily.cvd\|mirrors.dat" "noexit"
+ret=$(downloadDatas "$VARDIR/spool/clamav/" "clamav3" $randomize "clamav" "\|main.cvd\|bytecode.cvd\|daily.cvd\|mirrors.dat" "noexit")
 
 ## restart clamd daemon
-kill -USR2 `cat $VARDIR/run/clamav/clamd.pid 2>/dev/null` > /dev/null 2>&1 
-log "Database reloaded"
+if [[ "$ret" -eq "1" ]]; then
+	kill -USR2 `cat $VARDIR/run/clamav/clamd.pid 2>/dev/null` > /dev/null 2>&1 
+	log "Database reloaded"
+fi
 
 removeLockFile "$FILE_NAME"
